Extract shared helper functions to appropriate files
